处理了事务不能提交到数据库的问题，并将@Service与@Resource注解改用为@Named与@Inject注解

+++ src/main/java/com/dong/service/LinkApplication.java
@@ -10,6 +10,6 @@ public interface LinkApplication extends BaseApplication<LinkDTO, Long> {
 	
 	Page<LinkDTO> getPage(int currentPage, int pageSize);
 	
-	BigInteger getTotal();
+	Long getTotal();
 
 }
